requirements.txt
Django
djangorestframework
drf-spectacular
python-dotenv
pylint-django>=2.5.3
djangorestframework_simplejwt
django-filter
pytz